Refactor track selection helpers
